Merge new project after review. (#2)
* Delete InfixExpressionParser.java

removing all content from this branch to load my new version

* Delete input.txt

removing empty file

* Add files via upload

Uploading the entire revised project. Version is free of error and produces expected output. Prints output in a pleasing format to the console. Includes a Node, LinkedList, and Stack class, but created and coded by me. I did not use the built in versions.

* Update ExpressionEvaluator.java

Removal of an unnecessary comment.